requirements.txt
asgiref==3.4.1
Django==3.2.7
django-appconf==1.0.4
django-comments-dab==2.7.1
django-cors-headers==3.10.0
django-crispy-forms==1.12.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
Markdown==3.3.4
Pillow==8.3.2
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
